Create BookListCreator and move list creation logic there Remove from Crawler to BookListCreator: visitBookDetailsUrl and getBook methods and getCrawledBooks method list creation part
