+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/Crawler.java
@@ -2,7 +2,6 @@ package com.reka.lakatos.searchbyisbn.crawler.ervinszabolibrary;
 
 import com.reka.lakatos.searchbyisbn.crawler.BookCrawler;
 import com.reka.lakatos.searchbyisbn.document.Book;
-import com.reka.lakatos.searchbyisbn.exception.BookDownloadException;
 import com.reka.lakatos.searchbyisbn.exception.BookListDownloadException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,17 +12,15 @@ import org.springframework.stereotype.Component;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class Crawler implements BookCrawler {
 
-    private final BookCreator bookCreator;
     private final URLFactory urlFactory;
     private final PageReader pageReader;
+    private final BookListCreator bookListCreator;
 
     private static final String ISBN963 = "978963";
     private static final String ISBN615 = "978615";
@@ -53,25 +50,10 @@ public class Crawler implements BookCrawler {
     }
 
     private List<Book> getCrawledBooks() {
-        final Map<String, String> pageBooksInformation = getBooksDetailsInformation(
+        final Map<String, String> booksDetailsInformation = getBooksDetailsInformation(
                 urlFactory.createISBNSearchingUrl(page, searchingISBNMainGroup + isbnSeventhNumber, PAGE_SIZE));
 
-        return pageBooksInformation.entrySet().stream()
-                .map(bookInformationEntry -> urlFactory.createBookDetailsUrl(bookInformationEntry.getKey(),
-                        bookInformationEntry.getValue()))
-                .map(this::visitBookDetailsUrl)
-                .flatMap(Optional::stream)
-                .collect(Collectors.toList());
-    }
-
-    private Optional<Book> visitBookDetailsUrl(final String bookDetailsUrl) {
-        try {
-            return getBook(bookDetailsUrl);
-        } catch (Exception e) {
-            log.error("Exception happened while crawling book location: " + bookDetailsUrl, e);
-
-            return Optional.empty();
-        }
+        return bookListCreator.getBookListFromBooksDetailsInformation(booksDetailsInformation);
     }
 
     private Map<String, String> getBooksDetailsInformation(String ISBNSearchingUrl) {
@@ -81,17 +63,6 @@ public class Crawler implements BookCrawler {
             return pageReader.getBookDetailsLinkInformation(bookListPage);
         } catch (IOException e) {
             throw new BookListDownloadException("Unable to download the list book page! Url: " + ISBNSearchingUrl, e);
-        }
-    }
-
-    private Optional<Book> getBook(String bookDetailsUrl) {
-        try {
-            Document bookPage = Jsoup.connect(bookDetailsUrl).get();
-            return bookCreator.createBook(
-                    pageReader.getBookInformation(bookPage),
-                    PageReader.getSpecialSeparationCharacter());
-        } catch (IOException e) {
-            throw new BookDownloadException("Unable to download book page! Url: " + bookDetailsUrl, e);
         }
     }
 }
